Improve naming
* Avoid `RawXXXLayer` names that are meaningless, in favor of simple `XXXLayer`
* Rename original `XXXLayer` to `CumulativeXXXLayer`
