+++ plugins-CumulativeFloatLayers/build.sbt
@@ -1,0 +1,5 @@
+enablePlugins(Example)
+
+libraryDependencies += "com.thoughtworks.each" %% "each" % "3.3.1" % Test
+
+scalacOptions += "-Ypartial-unification"
