+++ plugins-FloatLayers/build.sbt
@@ -1,5 +1,3 @@
+incOptions in Test := incOptions.value.withRecompileOnMacroDef(true)
+
 enablePlugins(Example)
-
-libraryDependencies += "com.thoughtworks.each" %% "each" % "3.3.1" % Test
-
-scalacOptions += "-Ypartial-unification"
