+++ plugins-Operators/src/main/scala/com/thoughtworks/deeplearning/plugins/Operators.scala
@@ -4,7 +4,7 @@ import shapeless.{Poly1, Poly2}
 
 /** A plugin of definitions of polymorphic operators.
   *
-  * Those functions are implemented in [[RawFloatLayers]], [[RawDoubleLayers]] and [[RawINDArrayLayers]].
+  * Those functions are implemented in [[FloatLayers]], [[DoubleLayers]] and [[INDArrayLayers]].
   * @author 杨博 (Yang Bo)
   */
 trait Operators {
