--- plugins-RawFloatLayers/build.sbt
@@ -1,3 +1,0 @@
-incOptions in Test := incOptions.value.withRecompileOnMacroDef(true)
-
-enablePlugins(Example)
